Night commit
1. Global bans have been added to automatically ban child predators,
stalkers or any other serve offence in all the werewolf enforcer groups
when they join. You will be notified when this takes place in your group
and it can be appealed in Werewolf support which can be reached using
/support or /help.

2. A new command called /support has been added. This sends the user, or
replied user the link to our support group if they need help.

3. The help menu has been redesigned to be more helpful and take you
straight to support.

4. Kick has been modified to compensate for telegram ignoring unban
requests so this problem should be solved.

5. Warn has been rewritten to make it easier to work with.

6. Better logging abilities have been added throughout to make detecting
errors easier.